+++ NEWS.md
@@ -1,5 +1,7 @@
 # CHANGES IN servr VERSION 0.29
 
+- Added a function `httr()` to run R scripts and show output as HTML pages when serving a directory.
+
 - Added a new argument `response` to `httd()` to post-process the response.
 
 # CHANGES IN servr VERSION 0.28