+++ NEWS.md
@@ -1,3 +1,6 @@
+# CHANGES IN servr VERSION 0.30
+
+
 # CHANGES IN servr VERSION 0.29
 
 - Added a function `httr()` to run R scripts and show output as HTML pages when serving a directory.